nota de credito

+++ src/app/enterprise/sale/pages/createpresale/createpresale.component.ts
@@ -1,4 +1,4 @@
-import { Component, ElementRef, Input, OnInit, ViewChild } from '@angular/core';
+import { Component, ElementRef, HostListener, Input, OnInit, ViewChild } from '@angular/core';
 import { ProductSearchDto } from 'src/app/enterprise/product/model/dto/ProductSearchDto';
 import { ProductSearchService } from 'src/app/enterprise/product/service/productsearch.service';
 import { ResponseWsDto } from 'src/app/enterprise/shared/model/dto/ResponseWsDto';
@@ -20,6 +20,7 @@ import { PresaleDetailDto } from '../../model/dto/PresaleDetailDto';
 import { SaleDetailDto } from '../../model/dto/SaleDetailDto';
 import { ClientService } from '../../../client/service/client.service';
 import { ClientEntity } from '../../../client/model/entity/ClientEntity';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-createpresale',
@@ -35,7 +36,7 @@ export class CreatepresaleComponent implements OnInit {
   @ViewChild('cboDocumentType') cboDocumentType!: ElementRef<HTMLSelectElement>;
 
   productSearch : ProductSearchDto = new ProductSearchDto();
-  responsePageSearch : ResponsePageSearch = new ResponsePageSearch();
+  responsePageSearch : ResponsePageSearch<ProductSearchEntity> = new ResponsePageSearch();
   productList : ProductSearchEntity[] = [];
   productListHtml : ProductSearchEntity[][] = [];
   productInfoDtoSelect : ProductInfoDto = new ProductInfoDto();
@@ -53,7 +54,10 @@ export class CreatepresaleComponent implements OnInit {
   DocumentType : string = "";
   DocumentNum : string = "";
 
-  g_RptBusquedaProducto : RespuestaPaginacionDto = new RespuestaPaginacionDto();
+  lastKeypressTime: number = 0;
+  inputBuffer: string = ''; 
+
+  g_RptBusquedaProducto : RespuestaPaginacionDto<ProductSearchEntity> = new RespuestaPaginacionDto();
   g_ListaBoton : InfoPaginaDto[] = [];
 
 
@@ -74,29 +78,25 @@ export class CreatepresaleComponent implements OnInit {
     this.productSearch.Page = 1;
     this.findAllProduct();
     this.findDataForm(this.ShoppingCart.Headboard.PresaleCod);
-
   }
 
   ngOnInit(): void {
+    if(!this.ShoppingCart.Headboard.PresaleCod){
+      this.Clean();
+    }
     this.shoppingCartService.Init();
     this.updateShoppingCart();
   }
 
-  async findAllProduct()
-  {
-    const response : ResponseWsDto = await this.productSearchService.query(this.productSearch);
+  async InitNewSale(){
+    await this.createCode();
+  }
 
-    if( !response.ErrorStatus )
-    {
-        this.responsePageSearch = response.Data;
-        this.productList = this.responsePageSearch.resultSearch;
-
-        this.organizeProduct();
-
-        this.g_ListaBoton = [];
-        this.g_RptBusquedaProducto.addResultPage(this.responsePageSearch);
-        this.g_ListaBoton = this.CalcularPaginacion(this.g_RptBusquedaProducto);
-    }
+  async createCode(){
+    const response : ResponseWsDto = await this.presaleService.createCode();
+    
+    if(response.ErrorStatus) return;
+    this.ShoppingCart.Headboard.PresaleCod = String(response.Data);
   }
 
   async findDataForm(PresaleCod : string)
@@ -116,14 +116,57 @@ export class CreatepresaleComponent implements OnInit {
     }
   }
 
-  FiltrarProducto( p_num_pagina_busqueda : number  = 1)
+  async findAllProduct(IsBarcodeReaderInput : boolean = false)
+  {
+    const response : ResponseWsDto = await this.productSearchService.query(this.productSearch);
+
+    if( !response.ErrorStatus )
+    {
+        this.responsePageSearch = response.Data;
+        this.productList = this.responsePageSearch.resultSearch;
+
+        this.organizeProduct();
+
+        this.g_ListaBoton = [];
+        this.g_RptBusquedaProducto.addResultPage(this.responsePageSearch);
+        this.g_ListaBoton = this.CalcularPaginacion(this.g_RptBusquedaProducto);
+
+        if(IsBarcodeReaderInput){
+          this.addUnitDirect(this.productList[0].ProductCod);
+          this.txt_filtro_busqueda.nativeElement.value = "";
+        }
+    }
+
+    this.txt_filtro_busqueda.nativeElement.focus();
+  }
+
+  async addUnitDirect(ProductCod : string){
+    const response : ResponseWsDto = await this.productService.findDetailById(
+      ProductCod,this.session.getSessionStorageDto().StoreCod
+    );
+    const productInfoDto : ProductInfoDto = response.Data;
+
+    this.addUnit(productInfoDto,productInfoDto.VariantList[0]);
+  }
+
+  async subtractUnitDirect(ProductCod : string){
+    const response : ResponseWsDto = await this.productService.findDetailById(
+      ProductCod,this.session.getSessionStorageDto().StoreCod
+    );
+    const productInfoDto : ProductInfoDto = response.Data;
+
+    this.subtractUnit(productInfoDto,productInfoDto.VariantList[0]);
+  }
+  
+
+  FiltrarProducto( p_num_pagina_busqueda : number  = 1, IsBarcodeReaderInput : boolean = false)
   {
     if( p_num_pagina_busqueda <= 0 ) return;
 
     this.productSearch.StoreCod = this.session.getSessionStorageDto().StoreCod;
     this.productSearch.Query = this.txt_filtro_busqueda.nativeElement.value;
     this.productSearch.Page = p_num_pagina_busqueda;
-    this.findAllProduct();
+    this.findAllProduct(IsBarcodeReaderInput);
   }
 
   organizeProduct()
@@ -161,7 +204,7 @@ export class CreatepresaleComponent implements OnInit {
   }
 
 
-  CalcularPaginacion(P_InfoPaginacion : RespuestaPaginacionDto):InfoPaginaDto[]
+  CalcularPaginacion(P_InfoPaginacion : RespuestaPaginacionDto<ProductSearchEntity>):InfoPaginaDto[]
   {
       let l_numero_paginas : number[] = [];
       let l_ListaBoton : InfoPaginaDto[] = [];
@@ -272,15 +315,18 @@ export class CreatepresaleComponent implements OnInit {
 
   async findDetailById(ProductCod : string) 
   {
-    const response : ResponseWsDto = await this.productService.findDetailById({
-      ProductCod : ProductCod,
-      StoreCod : this.session.getSessionStorageDto().StoreCod
-    });
+    const response : ResponseWsDto = await this.productService.findDetailById(
+      ProductCod,this.session.getSessionStorageDto().StoreCod
+    );
 
     if( !response.ErrorStatus )
     {
         this.productInfoDtoSelect = response.Data;
         this.NumPhysicalStockTotal = this.productInfoDtoSelect.InfoList.map( item => item.NumPhysicalStock).reduce((a, b) => a + b, 0);
+    }
+
+    if(!this.ShoppingCart.Headboard.PresaleCod){
+      this.InitNewSale();
     }
   }
 
@@ -288,10 +334,10 @@ export class CreatepresaleComponent implements OnInit {
   {
     for(let Product of PresaleDetail.DetailList)
     {
-      const response : ResponseWsDto = await this.productService.findDetailById({
-        ProductCod : Product.ProductCod,
-        StoreCod : this.session.getSessionStorageDto().StoreCod
-      });
+      const response : ResponseWsDto = await this.productService.findDetailById(
+        Product.ProductCod,
+        this.session.getSessionStorageDto().StoreCod
+      );
 
       Product.ProductInfo = response.Data;
     }
@@ -308,8 +354,28 @@ export class CreatepresaleComponent implements OnInit {
 
   subtractUnit(ProductInfo : ProductInfoDto,ProductVariant : ProductVariantEntity)
   {
-    this.shoppingCartService.subtractUnit(ProductInfo,ProductVariant);
-    this.updateShoppingCart();
+    if(this.shoppingCartService.preventZeroSubtract(ProductInfo,ProductVariant)){
+      
+      Swal.fire({
+        title: '¿Estás seguro?',
+        text: 'Al modificar a 0 unidades el producto se eliminara del carrito',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Sí, eliminar',
+        cancelButtonText: 'No, cancelar'
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          this.shoppingCartService.subtractUnit(ProductInfo,ProductVariant);
+          this.updateShoppingCart();
+        }
+      });
+    }else{
+      this.shoppingCartService.subtractUnit(ProductInfo,ProductVariant);
+      this.updateShoppingCart();
+    }
+    
   }
 
   HandbookUnit(ProductInfo : ProductInfoDto,ProductVariant : ProductVariantEntity)
@@ -321,6 +387,7 @@ export class CreatepresaleComponent implements OnInit {
     }catch(e)
     {
       this.txt_NumUnit.nativeElement.value = "0";
+      this.updateShoppingCart();
     }
   }
 
@@ -341,8 +408,21 @@ export class CreatepresaleComponent implements OnInit {
 
   DeleteProduct(ProductCod : string)
   {
-    this.shoppingCartService.DeleteProduct(ProductCod);
-    this.updateShoppingCart();
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: 'El producto se eliminara del carrito',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, eliminar',
+      cancelButtonText: 'No, cancelar'
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        this.shoppingCartService.DeleteProduct(ProductCod);
+        this.updateShoppingCart();
+      }
+    });
   }
 
   async save()
@@ -443,6 +523,47 @@ export class CreatepresaleComponent implements OnInit {
     this.ShowClient = false;
     this.ShowClientRegister = false;
     this.ShowClientSearch = true;
+  }
+
+  FiltrarProductoEnter(event: KeyboardEvent){
+    const now = Date.now();
+    const timeDifference = now - this.lastKeypressTime;
+    let IsBarcodeReaderInput : boolean = false;
+
+    if (timeDifference < 50) {
+      this.inputBuffer += event.key;
+    } else {
+      this.inputBuffer = event.key;
+    }
+
+    this.lastKeypressTime = now;
+
+    if (event.key === 'Enter') {
+      if (this.isBarcodeScannerInput()) {
+        IsBarcodeReaderInput = true;
+        console.log('Entrada detectada como lector de código de barras:', this.inputBuffer);
+      } else {
+        IsBarcodeReaderInput = false;
+        console.log('Entrada detectada como manual:', this.inputBuffer);
+      }
+
+      this.FiltrarProducto(1,IsBarcodeReaderInput);
+      this.inputBuffer = '';
+    }
+  }
+
+  isBarcodeScannerInput(): boolean {
+    return this.inputBuffer.length > 5;
+  }
+
+  @HostListener('document:keydown', ['$event'])
+  onKeyDown(event: KeyboardEvent) {
+    if (event.altKey && (event.key === 'b' || event.key === 'B')) {
+      event.preventDefault();
+      window.scrollTo(0, 0);
+      this.txt_filtro_busqueda.nativeElement.focus();
+      this.txt_filtro_busqueda.nativeElement.select();
+    }
   }
 
 }
